+++ doc/whats-new.rst
@@ -20,6 +20,10 @@ New Features
 - ``compute=False`` is now supported by :py:meth:`DataTree.to_netcdf` and
   :py:meth:`DataTree.to_zarr`.
   By `Stephan Hoyer <https://github.com/shoyer>`_.
+- :py:func:`DataTree.from_dict` now supports passing in ``DataArray`` values,
+  and has a ``coords`` argument for specifying coordinates as ``DataArray``
+  objects (:pull:`10658`).
+  By `Stephan Hoyer <https://github.com/shoyer>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~